feat: add default search and save query to LS
